requirements.txt
ccxt>=4.4.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
aiofiles>=23.2.0
plotly>=5.17.0
streamlit>=1.28.0

# Dashboard dependencies
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-iconify>=0.1.2
flask>=2.3.0
gunicorn>=21.2.0

# Backtesting dependencies
backtrader>=1.9.78.123
zipline-reloaded>=3.0.4
ta>=0.10.2
yfinance>=0.2.18
